rename enum.py to myenum.py and update import
