+++ custom_components/midea_ac_lan/midea/backports/myenum.py
