+++ custom_components/midea_ac_lan/midea/core/device.py
@@ -2,7 +2,7 @@ import threading
 try:
     from enum import StrEnum
 except ImportError:
-    from ..backports.enum import StrEnum
+    from ..backports.myenum import StrEnum
 from enum import IntEnum
 from .security import LocalSecurity, MSGTYPE_HANDSHAKE_REQUEST, MSGTYPE_ENCRYPTED_REQUEST
 from .packet_builder import PacketBuilder
